+++ packages/analytics/analytics-chart/src/index.ts
@@ -1,8 +1,11 @@
 import AnalyticsChart from './components/AnalyticsChart.vue'
 import SimpleChart from './components/SimpleChart.vue'
 import SparklineChart from './components/SparklineChart.vue'
+import TopTalkersGrid from './components/top-talkers/TopTalkersGrid.vue'
+import TopTalkersColumn from './components/top-talkers/TopTalkersColumn.vue'
 
-export { AnalyticsChart, SimpleChart, SparklineChart }
+export { AnalyticsChart, SimpleChart, SparklineChart, TopTalkersGrid, TopTalkersColumn }
+export type { TopTalkersGridColumn } from './components/top-talkers/TopTalkersGrid.vue'
 
 /**
  * @deprecated Use dashboard-renderer for dashboard TopN tiles, or TableDataGrid
@@ -25,3 +28,5 @@ export { lookupStatusCodeColor } from './utils/customColors'
 export * from './utils/constants'
 export * from './utils/queryError'
 export * from './utils/scatter-adapters'
+export * from './utils/topn-columns'
+export { OTHER_DIMENSION_ID } from './constants'