+++ packages/core/sandbox-layout/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.6](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fsandbox-layout%402.7.5...%40kong-ui-public%2Fsandbox-layout%402.7.6) (2026-09-24)
+
+### Bug Fixes
+
+* **sandbox-layout:** make theme picker optional ([#3877](https://github.com/Kong/public-ui-components/issues/3877)) ([d394c0b](https://github.com/Kong/public-ui-components/commit/d394c0b3c4adc787e72300e431eae1a11221a260))
+
+
 ## [2.7.5](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fsandbox-layout%402.7.4...%40kong-ui-public%2Fsandbox-layout%402.7.5) (2026-09-21)
 
 **Note:** Version bump only for package @kong-ui-public/sandbox-layout