+++ packages/entities/entities-consumers/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.2](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-consumers%404.8.1...%40kong-ui-public%2Fentities-consumers%404.8.2) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumers
+
+
+
+
+
 ## [4.8.1](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-consumers%404.8.0...%40kong-ui-public%2Fentities-consumers%404.8.1) (2026-09-22)
 
 **Note:** Version bump only for package @kong-ui-public/entities-consumers